requirements.txt
Django==2.2.1
django-makemessages-xgettext==0.1.1
django-tinymce==2.8.0
django-widget-tweaks==1.4.3
Pillow==6.0.0
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
